// interfaces/ICharacter.ts
export interface ICharacter {
    id: number;
    name: string;
    status: string;
    species: string;
    gender: string;
    image: string;
    url: string;
}

// export interface Info {
//     count: number;
//     pages: number;
//     next: string;
//     prev?: any;
//   }
  
//   export interface Data {
//     info: Info,
//     results: Character[]
//   }
  